app_cv/requirements.txt
streamlit
langchain-core
langchain-community
langchain-cohere
langchain-chroma
pypdfium2-team
python-dotenv
requests
tenacity
cohere